Optimize removal of defs-, title-, desc-, metadata- and g-tags
The original code searched over all tags once of each type of these
tags and generated a list of matching tags (via
getElementsByTypeName).  However, some SVG files
(e.g. 1_42_polytope_7-cube.svg from #184) have so many nodes that this
takes up a significant amount of time.

As we do not really need a particular ordering between the tags, we
can optimize the removal by searching the DOM for all tag names at the
same time and evaluate them as we see them.  With this, we reduce the
DOM traversals from 5 to 1 for the concrete code snippet.

Runtime-wise, this seems to remove about 3-4 minutes in python3.

Signed-off-by: Niels Thykier <niels@thykier.net>